// apps/web/src/lib/api/generated/model/badgeResponseTitle.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Petclinic MSS301 — Aggregated API
 * Aggregated from: auth, customers, vets, visits
 * OpenAPI spec version: 1.0.0
 */

export type BadgeResponseTitle = typeof BadgeResponseTitle[keyof typeof BadgeResponseTitle];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BadgeResponseTitle = {
  ROOKIE: 'ROOKIE',
  EXPERIENCED: 'EXPERIENCED',
  MASTER: 'MASTER',
  SURGERY_EXPERT: 'SURGERY_EXPERT',
  RESEARCH_AWARD: 'RESEARCH_AWARD',
  TOP_RATED: 'TOP_RATED',
} as const;